fix call observer
